generalize SplitSimulatorWindow
