+++ lib/gui/simulator.py
@@ -1085,6 +1085,65 @@ class SimulatorWindow(wx.Frame):
         self.parent.attach_simulator()
 
 
+class SplitSimulatorWindow(wx.Frame):
+    def __init__(self, panel_class, **kwargs):
+        super().__init__(None, title=_("Embroidery Params"))
+
+        self.detached_simulator_frame = None
+        self.splitter = wx.SplitterWindow(self)
+        self.simulator_panel = SimulatorPanel(self.splitter, detach_callback=self.toggle_detach_simulator)
+        self.settings_panel = panel_class(self.splitter, simulator=self.simulator_panel, **kwargs)
+
+        self.splitter.SplitVertically(self.settings_panel, self.simulator_panel)
+        self.splitter.SetMinimumPaneSize(100)
+
+        icon = wx.Icon(os.path.join(get_resource_dir("icons"), "inkstitch256x256.png"))
+        self.SetIcon(icon)
+
+        self.sizer = wx.BoxSizer(wx.VERTICAL)
+        self.sizer.Add(self.splitter, 1, wx.EXPAND)
+        self.SetSizer(self.sizer)
+
+        self.SetSizeHints(self.sizer.CalcMin())
+        self.Maximize()
+
+        self.Bind(wx.EVT_CLOSE, self.on_close)
+
+    def on_close(self, event):
+        if self.detached_simulator_frame:
+            self.detached_simulator_frame.Destroy()
+        self.Destroy()
+
+    def toggle_detach_simulator(self):
+        if self.detached_simulator_frame:
+            self.attach_simulator()
+        else:
+            self.detach_simulator()
+
+    def attach_simulator(self):
+        self.detached_simulator_frame.detach_simulator_panel()
+        self.simulator_panel.Reparent(self.splitter)
+        self.splitter.SplitVertically(self.settings_panel, self.simulator_panel)
+        self.detached_simulator_frame.Destroy()
+        self.detached_simulator_frame = None
+        self.Maximize()
+        self.splitter.UpdateSize()
+        self.SetFocus()
+        self.Raise()
+
+    def detach_simulator(self):
+        simulator_rect = self.simulator_panel.GetScreenRect()
+        settings_rect = self.settings_panel.GetScreenRect()
+        self.splitter.Unsplit()
+        self.detached_simulator_frame = SimulatorWindow(panel=self.simulator_panel, parent=self)
+        self.detached_simulator_frame.Show()
+        self.splitter.SetMinimumPaneSize(100)
+        self.Maximize(False)
+        self.SetMinSize((200, 200))
+        self.SetClientRect(settings_rect)
+        self.detached_simulator_frame.SetClientRect(simulator_rect)
+
+
 class PreviewRenderer(Thread):
     """Render stitch plan in a background thread."""
